pyproject.toml
[project]
name = "swelab"
version = "0.1.0"
description = "NO2 Satellite Data Downscaling using AI/ML"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "rasterio>=1.4.3",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.42.0",
    "scipy>=1.12.0",
    "xgboost>=2.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools]
py-modules = ["main", "model", "database", "utils"]